migrate AllowTokens to typescript
