+++ federator/src/contracts/AllowTokensFactory.ts
@@ -2,8 +2,8 @@ import abiAllowTokensV0 from '../../../bridge/abi/AllowTokensV0.json';
 import abiAllowTokensV1 from '../../../bridge/abi/AllowTokensV1.json';
 import abiAllowTokensV2 from '../../../bridge/abi/AllowTokens.json';
 import abiBridgeV3 from '../../../bridge/abi/BridgeV3.json';
-import IAllowTokensV1 from './IAllowTokensV1';
-import IAllowTokensV0 from './IAllowTokensV0';
+import { IAllowTokensV1 } from './IAllowTokensV1';
+import { IAllowTokensV0 } from './IAllowTokensV0';
 import CustomError from '../lib/CustomError';
 import utils from '../lib/utils';
 import { Config } from '../../config/types';
@@ -48,7 +48,7 @@ export class AllowTokensFactory extends ContractFactory {
         return new IAllowTokensV1(allowTokensContract);
       case V0:
         allowTokensContract = this.getContractByAbi(abiAllowTokensV0 as AbiItem[], address, web3);
-        return new IAllowTokensV0(allowTokensContract);
+        return new IAllowTokensV0(allowTokensContract, chainId);
       default:
         throw Error('Unknown AllowTokens contract version');
     }
