--- federator/src/contracts/IAllowTokensV0.js
@@ -1,38 +1,0 @@
-module.exports = class IAllowTokensV0 {
-    constructor(allowTokensContract, chainId) {
-        this.allowTokensContract = allowTokensContract;
-        this.mapTokenInfoAndLimits = {};
-        this.chainId = chainId;
-    }
-
-    getVersion() {
-        return 'v0';
-    }
-
-    async getConfirmations() {
-        let confirmations = 0; //for rsk regtest and ganache
-        if (this.chainId == 31 || this.chainId == 42) { // rsk testnet and kovan
-            confirmations = 10
-        }
-        if (this.chainId == 1) { //ethereum mainnet 24hs
-            confirmations = 240
-        }
-        if (this.chainId == 30) { // rsk mainnet 24hs
-            confirmations = 120
-        }
-        return {
-            smallAmountConfirmations: confirmations,
-            mediumAmountConfirmations: confirmations,
-            largeAmountConfirmations: confirmations,
-        }
-    }
-
-
-    async getLimits(tokenAddress) {
-        return {
-            allowed: true,
-            mediumAmount: -1,
-            largeAmount: 0,
-        };
-    }
-}
